requirements.txt
matplotlib==3.1.2
numpy==1.16.4
pandas==0.24.2
seaborn==0.9.0
requests==2.22.0
omdb==0.10.1
Pillow==6.2.1
Flask==1.1.1